Move body generator for adjacently tagged enums to its own module
Cut-paste
